Move datamaps to separate package
This is to avoid pulling in all the CLI dependencies when using this
