+++ .air.toml
@@ -16,7 +16,7 @@ exclude_regex = [
   "_templ\\.go",
   "\\.sql\\.go",
   ".*models/(models|copyfrom|db).go",
-  "cmd/datamaps",
+  "datamaps",
 ]
 exclude_unchanged = false
 follow_symlink = false
